:sparkles: feat: user able to delete a comment as post author and comment author

+++ backend/src/business/post-business.ts
@@ -38,6 +38,12 @@ export class POST_BUSINESS {
         const comment = await this.postDatabase.getCommentById( commentId )
         if ( !comment ) throw new CustomError( 404, "Comentário não encontrado" )
 
+        //vericar se o user é o autor do post ou do comentário
+
+        const post = await this.postDatabase.getPostById( comment.post_id )
+        if ( !post ) throw new CustomError( 404, "Post não encontrado" )
+
+        if ( post.author_id !== userId ) throw new CustomError( 401, "Você não tem permissão para deletar este comentário" )
         if ( comment.author_id !== userId ) throw new CustomError( 401, "Você não tem permissão para deletar este comentário" )
 
         await this.postDatabase.deleteComment( commentId )
